[proto-plus-python] Fix Fuzz Target for Python 3.10
In Python 3.10, the fuzz target fails if the `FuzzMsg` class is defined
within `TestOneInput` with the exception:
```
TypeError: Couldn't build proto file into descriptor pool: duplicate
file name __main____default_package.fuzzmsg.proto
```

This was not an issue in Python 3.8. This change is necessary to remain
compatible with the Python version used in the base-builder and
base-runner images.