[pillow] fix build failure after Python 10.x upgrade
Prior to the Python 10 upgrade in the base-runner image there was
no version-less `python` executable; only `python3` was available
so some projects worked around that by symlinking the latter to the
former.

The Python version upgrade also included a change that applied similar
symbolic links of common executable names for convenience. This caused
a conflict for projects applying it again, and broke the build.

Removing the now redundant python3 -> python executable symlink resolves
the failure.